Fixed clock scaling issue
If the clock size is set larger than the Budgie Panel size, it will draw the clock to fit to the panel size.  However, if you enlarge the Budgie Panel, the clock size will now increase as well until it reaches the chosen clock size (while not exceeding the size of the Budgie Panel)

Essentially, the clock will now shrink and grow as the Budgie Panel size is resized, until it reaches the selected size.